feat: create HeroSection and IdentityView components; update routing and remove obsolete views
