+++ src/components/features/identity/HeroSection.vue
@@ -1,0 +1,27 @@
+<template>
+    <section class=" h-screen grid lg:grid-cols-2 overflow-hidden items-center">
+        <!-- Kolom Kiri - Content -->
+        <div class="px-8 lg:px-16 space-y-6">
+            <!-- Title -->
+            <h1 class="text-6xl font-semibold leading-tight">
+                Buku Digital<br>
+                Teknik Komputer
+            </h1>
+
+            <!-- Description -->
+            <p class="text-xl font-normal text-gray-600 leading-relaxed">
+                Lorem ipsum dolor sit amet, consectetur adipiscing elit.
+                Sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.
+            </p>
+        </div>
+
+        <!-- Kolom Kanan -->
+        <div class="relative h-screen">
+            <img src="@/assets/images/hero-bg.png" alt="Hero Image" class="w-full h-full object-cover" />
+        </div>
+
+    </section>
+</template>
+
+<script setup>
+</script>
