--- src/views/AboutView.vue
@@ -1,21 +1,0 @@
-<template>
-  <div class="about">
-    <h1>About Page</h1>
-    <p>This is the about page - placeholder content</p>
-    
-    <div class="mt-4">
-      <RouterLink to="/" class="text-blue-500 hover:underline">Back to Home</RouterLink>
-    </div>
-  </div>
-</template>
-
-<script setup>
-// About page logic will go here
-</script>
-
-<style scoped>
-.about {
-  padding: 2rem;
-  text-align: center;
-}
-</style>
