+++ src/views/IdentityView.vue
@@ -1,0 +1,10 @@
+<template>
+  <div class="identity-view">
+    <HeroSection />
+  </div>
+
+</template>
+
+<script setup>
+import HeroSection from '@/components/features/identity/HeroSection.vue';
+</script>
